requirements.txt
google-ai-generativelanguage==0.6.6
streamlit==1.36.0
flask==3.0.3
flask-cors==4.0.1